Gizmos v1.6
Ref: https://github.com/friction2d/friction/pull/601
