+++ src/core/canvasselectedboxesactions.cpp
@@ -412,18 +412,17 @@ void Canvas::scaleSelectedBy(const qreal scaleXBy,
 void Canvas::shearSelectedBy(const qreal shearXBy,
                              const qreal shearYBy,
                              const QPointF &absOrigin,
-                             const bool startTrans) {
-    if(mSelectedBoxes.isEmpty()) return;
-    if(mDocument.fLocalPivot) {
+                             const bool startTrans)
+{
+    if (mSelectedBoxes.isEmpty()) { return; }
+    if (mDocument.fLocalPivot) {
         for(const auto &box : mSelectedBoxes) {
-            if(startTrans) {
-                box->startShearTransform();
-            }
+            if (startTrans) { box->startShearTransform(); }
             box->shear(shearXBy, shearYBy);
         }
     } else {
         for(const auto &box : mSelectedBoxes) {
-            if(startTrans) {
+            if (startTrans) {
                 box->startShearTransform();
                 box->startPosTransform();
                 box->saveTransformPivotAbsPos(absOrigin);
