+++ src/core/canvasselectedpointsactions.cpp
@@ -358,24 +358,25 @@ void Canvas::scaleSelectedPointsBy(const qreal scaleXBy,
 void Canvas::shearSelectedPointsBy(const qreal shearXBy,
                                    const qreal shearYBy,
                                    const QPointF &absOrigin,
-                                   const bool startTrans) {
-    if(mSelectedPoints_d.isEmpty()) return;
-    if(startTrans) {
-        if(mDocument.fLocalPivot) {
-            for(const auto& point : mSelectedPoints_d) {
+                                   const bool startTrans)
+{
+    if (mSelectedPoints_d.isEmpty()) { return; }
+    if (startTrans) {
+        if (mDocument.fLocalPivot) {
+            for (const auto& point : mSelectedPoints_d) {
                 point->startTransform();
                 point->saveTransformPivotAbsPos(point->getAbsolutePos());
                 point->shearRelativeToSavedPivot(shearXBy, shearYBy);
             }
         } else {
-            for(const auto& point : mSelectedPoints_d) {
+            for (const auto& point : mSelectedPoints_d) {
                 point->startTransform();
                 point->saveTransformPivotAbsPos(absOrigin);
                 point->shearRelativeToSavedPivot(shearXBy, shearYBy);
             }
         }
     } else {
-        for(const auto& point : mSelectedPoints_d) {
+        for (const auto& point : mSelectedPoints_d) {
             point->shearRelativeToSavedPivot(shearXBy, shearYBy);
         }
     }
